return a definitive answer on GetParameters
Windows relies on this call. Avoid the "FAILED" response and indicate T=1 block transmission as selected mode in any case the parameters are accessed. T=1 is the preferable way to go. If you change the ATR of the card to emulate to T=0-only then adjust the ProtocolNum here to ensure correct protocol behaviour on the host operating system.